==> 1-DNAseq/8-poolDseq-5cnt_2EY_ct-TSS_Bar_sumCnt-mtx.py <==
### python 3 codes ###
# remotely use Pitt HTC cluster
# python=python/anaconda3.8-2020.11 (Python 3.8.5.final.0); pandas=1.1.3; numpy=1.19.2;

### make pooled DNAseq TSS-Bar-DNA_reads_count matrix
#   1. keep TSS-Barcode variants that
#      (a) >=5 reads in libraries
#      (b) exist in >=2 Ecoli+3Yreps libraries
#   2. pool DNAseq reads with sum of reads from Ecoli+3Yreps libraries

# Input-1s: [*-CXed-BarFltr-Tss_Bar_Cnt-Mtx.txt] generated by <7-BarMltTss_fltr.py>, matrix after one_Barcode vs multiple_TSSs filter

# Output-1s: [*-poolDs-5cnt_2EY_ct-TSS_Bar_EYCnt-mtx.txt], matrix containing kept TSS-Bar and counts in each Ec/Y libraries
    # 6 columns, sep='\t': [TSS] - [Barcode_Ds] - [Ec] - [rep1] - [rep2] - [rep3]

import pandas as pd
from functools import reduce

cmn_folder = '/bgfs/ckaplan/Yunye/0-common_files/'
inp_folder = '/bgfs/ckaplan/Yunye/3-TSS_sequence_library/1-WT-DNAseq-TxGen_18179Kap/7-BarMltTss_fltr/'
out_folder = '/bgfs/ckaplan/Yunye/3-TSS_sequence_library/1-WT-DNAseq-TxGen_18179Kap/8-poolDseq-5cnt_2EY_ct/'

smp_info = pd.read_csv(cmn_folder+'DTmerge-MASTER-all_final-info_table.csv', na_filter= False) # w/o filling empty cells as NaN
# to only process samples with DNAseq:
smp_info = smp_info[smp_info['super_id']<=12]

for lib in ['AYR','BYR','ARY']:
    dfs = list()
    for _, smp in smp_info[smp_info['lib']==lib].iterrows():
        file_prefix = smp['lib']+'_'+smp['PolII']+'_'+smp['rep']+'-Ds-'
        temp_df = pd.read_csv(inp_folder+file_prefix+'CXed-BarFltr-Tss_Bar_Cnt-Mtx.txt', sep='\t', index_col=False)
        # only keep TSS-bar variants that have >=5 reads:
        temp_df = temp_df[temp_df['DNA_Count']>=5]
        dfs.append(temp_df.rename(columns={'DNA_Count':smp['rep']}))
    mtx = reduce(lambda left, right: pd.merge(left, right, on=['TSS', 'Barcode_Ds'], how = 'outer'), dfs)
    
    # only keep TSS-bar variants that exist in >=2 Ecoli+3Yreps libraries:
    mtx = mtx[(4-mtx[['Ec','rep1','rep2','rep3']].isnull().sum(axis=1)) >=2]
    mtx.to_csv(out_folder+smp['lib']+'-poolDs-5cnt_2EY_ct-TSS_Bar_EYCnt-mtx.txt', sep='\t', index=False, header=True)
